Tower of Hanoi in C
The Tower of Hanoi is a mathematical game or puzzle. It consists of
three rods, and a number of disks of different sizes which can slide
onto any rod. The puzzle starts with the disks in a neat stack in
ascending order of size on one rod, the smallest at the top, thus
making a conical shape.